Implemented saving/loading/deleting of games in form2.
Some testdata in database.

Added methods SetLoadedGame and GetCurrentGame in Form1, to set a loaded game to currentGame in gameLogicInstance and to save currentGame in gameLogicInstance.

Changed Repository slightly.
Changed name of GameId in Generation to Game for database access reasons.

Added method SetLoadedGame to GameLogic to update currentGame to the loaded one and currentBoad to the first generation of currentGame.